T2D_Chatbot_for_HR/requirements.txt
requests>=2.31.0
numpy>=1.24.0
scikit-learn>=1.3.0
transformers>=4.41.0
tqdm>=4.67.1
# Astro Runtime includes the following pre-installed providers packages: https://www.astronomer.io/docs/astro/runtime-image-architecture#provider-packages

apache-airflow[postgres]==2.8.1
pandas
openpyxl
pandas==2.0.3
openpyxl==3.1.5
pdfplumber==0.11.5
pypdf==5.9.0
sentence-transformers==3.2.1
faiss-cpu==1.8.0
torch>=2.1
